+++ src/main/java/com/example/draugar/Hnit.java
@@ -1,5 +1,5 @@
 package com.example.draugar;
 
 public interface Hnit {
-    double[] Hnit();
+    int[] Hnit();
 }
