+++ src/test/java/com/example/draugar/ClaydeTest.java
@@ -6,10 +6,10 @@ import org.junit.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ClaydeTest {
-    private final double[] A = {0, 0};
-    private final double[] B = {10, 10};
-    private final double[] HOME = {6, 6};
-    private final double[] HOME_BASE = {0, 10};
+    private final int[] A = {0, 0};
+    private final int[] B = {10, 10};
+    private final int[] HOME = {6, 6};
+    private final int[] HOME_BASE = {0, 10};
     private Clyde clyde;
     private Pacman pacman;
 
@@ -25,16 +25,16 @@ public class ClaydeTest {
 
     @Test
     public void testDrugaRikniritEltaUndanRadius() {
-        double[] a = {2, 5};
-        assertEquals(89, (int) clyde.drauaReikniritd(a));
+        int[] a = {2, 5};
+        assertEquals(89, clyde.drauaReikniritd(a));
 
-        double[] b = {1, 6};
+        int[] b = {1, 6};
         assertEquals(97, (int) clyde.drauaReikniritd(b));
 
-        double[] c = {2, 7};
+        int[] c = {2, 7};
         assertEquals(73, (int) clyde.drauaReikniritd(c));
 
-        double[] d = {3, 6};
+        int[] d = {3, 6};
         assertEquals(65, (int) clyde.drauaReikniritd(d));
     }
 
@@ -43,17 +43,17 @@ public class ClaydeTest {
         clyde.setCenterX(8);
         clyde.setCenterY(8);
 
-        double[] a = {8, 7};
-        assertEquals(73, (int) clyde.drauaReikniritd(a));
+        int[] a = {8, 7};
+        assertEquals(73,  clyde.drauaReikniritd(a));
 
-        double[] b = {7, 8};
-        assertEquals(53, (int) clyde.drauaReikniritd(b));
+        int[] b = {7, 8};
+        assertEquals(53,  clyde.drauaReikniritd(b));
 
-        double[] c = {8, 9};
-        assertEquals(65, (int) clyde.drauaReikniritd(c));
+        int[] c = {8, 9};
+        assertEquals(65,  clyde.drauaReikniritd(c));
 
-        double[] d = {9, 8};
-        assertEquals(85, (int) clyde.drauaReikniritd(d));
+        int[] d = {9, 8};
+        assertEquals(85,  clyde.drauaReikniritd(d));
     }
 
 
@@ -61,34 +61,34 @@ public class ClaydeTest {
     public void testDrugaRikniritNotEltaUndanRadius() {
         clyde.setElta(false);
 
-        double[] a = {2, 5};
-        assertEquals(29, (int) clyde.drauaReikniritd(a));
+        int[] a = {2, 5};
+        assertEquals(29,  clyde.drauaReikniritd(a));
 
-        double[] b = {1, 6};
-        assertEquals(17, (int) clyde.drauaReikniritd(b));
+        int[] b = {1, 6};
+        assertEquals(17,  clyde.drauaReikniritd(b));
 
-        double[] c = {2, 7};
-        assertEquals(13, (int) clyde.drauaReikniritd(c));
+        int[] c = {2, 7};
+        assertEquals(13,  clyde.drauaReikniritd(c));
 
-        double[] d = {3, 6};
-        assertEquals(25, (int) clyde.drauaReikniritd(d));
+        int[] d = {3, 6};
+        assertEquals(25,  clyde.drauaReikniritd(d));
     }
 
     @Test
     public void testDrugaRikniritEaten() {
         clyde.setEtan(true);
 
-        double[] a = {6, 5};
-        assertEquals(1, (int) clyde.drauaReikniritd(a));
+        int[] a = {6, 5};
+        assertEquals(1,  clyde.drauaReikniritd(a));
 
-        double[] b = {5, 6};
-        assertEquals(1, (int) clyde.drauaReikniritd(b));
+        int[] b = {5, 6};
+        assertEquals(1,  clyde.drauaReikniritd(b));
 
-        double[] c = {6, 7};
-        assertEquals(1, (int) clyde.drauaReikniritd(c));
+        int[] c = {6, 7};
+        assertEquals(1,  clyde.drauaReikniritd(c));
 
-        double[] d = {7, 6};
-        assertEquals(1, (int) clyde.drauaReikniritd(d));
+        int[] d = {7, 6};
+        assertEquals(1,  clyde.drauaReikniritd(d));
     }
 
     @Test
